+++ talkTek/Controller/Home/Tabs/HotViewController.swift
@@ -139,7 +139,7 @@ extension HotViewController {
         homeCourses.priceOnSales = dictionary["priceOnSales"] as? Int
         homeCourses.priceOrigin = dictionary["priceOrigin"] as? Int
         homeCourses.scorePeople = dictionary["scorePeople"] as? Int
-        homeCourses.scoreTotal = dictionary["scoreTotal"] as? Int
+        homeCourses.scoreTotal = dictionary["scoreTotal"] as? Double
         homeCourses.studentNumber = dictionary["studentNumber"] as? Int
         homeCourses.tags = dictionary["tags"] as! [String]
